use f# for transpiling and refactor cli
