+++ KubeOps.sln
@@ -38,7 +38,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KubeOps.Operator.Web", "src
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KubeOps.Operator.Web.Test", "test\KubeOps.Operator.Web.Test\KubeOps.Operator.Web.Test.csproj", "{BC6E6D3C-6404-4B01-9D72-AA16FEEBFF5C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KubeOps.Transpiler", "src\KubeOps.Transpiler\KubeOps.Transpiler.csproj", "{9C155732-5CE8-4F1D-8D5F-AB60DB1A1B5C}"
+Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "KubeOps.Transpiler", "src\KubeOps.Transpiler\KubeOps.Transpiler.fsproj", "{E3C31355-20F8-46C9-81E8-7BBF85CE7459}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KubeOps.Transpiler.Test", "test\KubeOps.Transpiler.Test\KubeOps.Transpiler.Test.csproj", "{47914451-147D-427E-B150-9C47DBF28F2C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -59,7 +61,8 @@ Global
 		{162EE2A9-F8BA-4D52-9D65-67CCDB76EAC9} = {C587731F-8191-4A19-8662-B89A60FE79A1}
 		{E5012C55-B34E-4672-AF4C-C233B45E8EA2} = {4DB01062-6DC5-4028-BB72-C0619C2F5F2E}
 		{BC6E6D3C-6404-4B01-9D72-AA16FEEBFF5C} = {C587731F-8191-4A19-8662-B89A60FE79A1}
-		{9C155732-5CE8-4F1D-8D5F-AB60DB1A1B5C} = {4DB01062-6DC5-4028-BB72-C0619C2F5F2E}
+		{E3C31355-20F8-46C9-81E8-7BBF85CE7459} = {4DB01062-6DC5-4028-BB72-C0619C2F5F2E}
+		{47914451-147D-427E-B150-9C47DBF28F2C} = {C587731F-8191-4A19-8662-B89A60FE79A1}
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{E9A0B04E-D90E-4B94-90E0-DD3666B098FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -102,9 +105,13 @@ Global
 		{BC6E6D3C-6404-4B01-9D72-AA16FEEBFF5C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BC6E6D3C-6404-4B01-9D72-AA16FEEBFF5C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BC6E6D3C-6404-4B01-9D72-AA16FEEBFF5C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9C155732-5CE8-4F1D-8D5F-AB60DB1A1B5C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9C155732-5CE8-4F1D-8D5F-AB60DB1A1B5C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9C155732-5CE8-4F1D-8D5F-AB60DB1A1B5C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9C155732-5CE8-4F1D-8D5F-AB60DB1A1B5C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E3C31355-20F8-46C9-81E8-7BBF85CE7459}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E3C31355-20F8-46C9-81E8-7BBF85CE7459}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E3C31355-20F8-46C9-81E8-7BBF85CE7459}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E3C31355-20F8-46C9-81E8-7BBF85CE7459}.Release|Any CPU.Build.0 = Release|Any CPU
+		{47914451-147D-427E-B150-9C47DBF28F2C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{47914451-147D-427E-B150-9C47DBF28F2C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{47914451-147D-427E-B150-9C47DBF28F2C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{47914451-147D-427E-B150-9C47DBF28F2C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
