+++ src/KubeOps.Abstractions/KubeOps.Abstractions.csproj
@@ -20,4 +20,5 @@
     </ItemGroup>
 
 
+
 </Project>
