+++ src/KubeOps.Abstractions/Rbac/RbacAttribute.cs
@@ -1,0 +1,8 @@
+﻿namespace KubeOps.Abstractions.Rbac;
+
+/// <summary>
+/// Abstract base class for all RBAC attributes.
+/// </summary>
+public abstract class RbacAttribute : Attribute
+{
+}
