+++ src/KubeOps.Cli/Commands/Generator/CrdGenerator.cs
@@ -1,6 +1,5 @@
 ﻿using KubeOps.Cli.Output;
 using KubeOps.Cli.SyntaxObjects;
-using KubeOps.Transpiler.Crds;
 
 using McMaster.Extensions.CommandLineUtils;
 
@@ -53,8 +52,8 @@ internal class CrdGenerator
 
         _output.WriteLine($"Generate CRDs from project: {projectFile}.");
 
-        var parser = await ProjectParser.CreateAsync(projectFile);
-        await foreach (var crd in parser.Entities().Select(CrdTranspiler.Create))
+        var parser = new ProjectParser(projectFile);
+        await foreach (var crd in parser.Entities().Select(Transpiler.Crds.Transpile))
         {
             _result.Add($"{crd.Metadata.Name.Replace('.', '_')}.{Format.ToString().ToLowerInvariant()}", crd);
         }
