+++ src/KubeOps.Cli/Commands/Generator/Generator.cs
@@ -4,11 +4,7 @@ namespace KubeOps.Cli.Commands.Generator;
 
 [Command("generator", "gen", "g", Description = "Generates elements related to an operator. (Aliases: gen, g)")]
 [Subcommand(typeof(CrdGenerator))]
-
-// [Subcommand(typeof(DockerGenerator))]
-// [Subcommand(typeof(InstallerGenerator))]
-// [Subcommand(typeof(OperatorGenerator))]
-// [Subcommand(typeof(RbacGenerator))]
+[Subcommand(typeof(RbacGenerator))]
 internal class Generator
 {
     public int OnExecute(CommandLineApplication app)
