+++ src/KubeOps.Cli/KubeOps.Cli.csproj
@@ -30,6 +30,6 @@
 
     <ItemGroup>
         <ProjectReference Include="..\KubeOps.Abstractions\KubeOps.Abstractions.csproj"/>
-        <ProjectReference Include="..\KubeOps.Transpiler\KubeOps.Transpiler.csproj" />
+        <ProjectReference Include="..\KubeOps.Transpiler\KubeOps.Transpiler.fsproj" />
     </ItemGroup>
 </Project>
