--- src/KubeOps.Transpiler/Crds/CrdConversionException.cs
@@ -1,18 +1,0 @@
-namespace KubeOps.Transpiler.Crds;
-
-public class CrdConversionException : Exception
-{
-    public CrdConversionException()
-    {
-    }
-
-    public CrdConversionException(string message)
-        : base(message)
-    {
-    }
-
-    public CrdConversionException(string message, Exception innerException)
-        : base(message, innerException)
-    {
-    }
-}
