--- src/KubeOps.Transpiler/Crds/CrdPropertyTypeException.cs
@@ -1,19 +1,0 @@
-namespace KubeOps.Transpiler.Crds;
-
-public class CrdPropertyTypeException : Exception
-{
-    public CrdPropertyTypeException()
-        : this("The given property has an invalid type.")
-    {
-    }
-
-    public CrdPropertyTypeException(string message)
-        : base(message)
-    {
-    }
-
-    public CrdPropertyTypeException(string message, Exception innerException)
-        : base(message, innerException)
-    {
-    }
-}
