+++ src/KubeOps.Transpiler/README.md
@@ -1,1 +1,1 @@
-transpiler :)
+transpiler
