+++ test/KubeOps.Transpiler.Test/UnitTest1.cs
@@ -1,0 +1,9 @@
+namespace KubeOps.Transpiler.Test;
+
+public class UnitTest1
+{
+    [Fact]
+    public void Test1()
+    {
+    }
+}
